# Conflicts:
#	tests/integration/cases/literalize_ref_in_dict/Norg_ref@v0_14.norg
#	tests/integration/cases/literalize_ref_in_mixed_list/Norg_ref@v0_14.norg

+++ tests/integration/cases/list_beyond_i64/Odin@dev_2024.odin
@@ -4,7 +4,7 @@ package main
 main :: proc() {
 my_data := [dynamic]any{
 	9223372036854775807,
-	9223372036854775808,
+	u64(9223372036854775808),
 }
 _ = my_data
 }
