+++ tests/integration/cases/record_wide_int/Odin_heterogeneous_strategy_record_integer_hex@dev_2024.odin
@@ -5,7 +5,7 @@ Record0 :: struct { quantity: int, big: u64, ratio: f64, label: string, ok: bool
 main :: proc() {
 my_data := Record0{
 	quantity = 0xf4240,
-	big = 0xffffffffffffffff,
+	big = u64(0xffffffffffffffff),
 	ratio = 2.5,
 	label = "tag",
 	ok = true,
